Remove singularity and docker files without subclassing
